luna/requirements.txt
discord.py[voice]
PyNaCl==1.5.0
dotenv
